Remove Bad Chrome Interactions with Native Objects.
This removes the Chrome interactions in the Dev Tools Command Line that
use native functions that are now wrapped.

The issue is that Chrome will use functions behind the scenes that are
unsupported by your target browser selection and you will get spammed
with false data that isn’t from your code.